% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{An R function with nifty HTML documentation}
\usage{
hello()
}
\value{
Prints a value
}
\description{
Some stuff about the function
}
\details{
\feedbackfooter{'noamross/htmlhelp'}{'R/hello.R'}{TRUE}
}

